Experiment with passing 'self' as argument to each method.
This provides about 5% speedier method dispatch.
But it fails test 70-finalize.
Also costs new (magnitudes higher) instance construction time proportional to number of instances already in the system.
Both symptoms are because a reference to the instance is captured in a static variable in the instance dispatcher.
That prevents gc from ever collecting the instance.
Jim offers no weak reference feature to cope with that.